almost
Hud, sort of almost renders how i want it, but in an attempt to get it EXACTLY where I want it, it disappears and defies expectations.

I got the "tap" to work, but fixed TargetImages work fine that are called in whilst editing.

The ones that are generated at run-time seam to be stuck to the world, and refuse to track as expected.